+++ src/features/downloads/Downloads.constants.ts
@@ -6,6 +6,8 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
+import { d } from 'koration';
+
 export const DOWNLOAD_AHEAD = {
     min: 2,
     max: 10,
@@ -17,4 +19,16 @@ export const DOWNLOAD_CONVERSION_COMPRESSION = {
     min: 0,
     max: 1,
     step: 0.01,
+};
+
+export const IMAGE_CONVERSION_CALL_TIMEOUT = {
+    min: d(10).seconds.inWholeSeconds,
+    max: d(10).minutes.inWholeSeconds,
+    step: d(10).seconds.inWholeSeconds,
+};
+
+export const IMAGE_CONVERSION_CONNECT_TIMEOUT = {
+    min: d(10).seconds.inWholeSeconds,
+    max: d(10).minutes.inWholeSeconds,
+    step: d(10).seconds.inWholeSeconds,
 };